% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.sequences.R
\name{check_lengths}
\alias{check_lengths}
\title{check_lengths}
\usage{
check_lengths(seqlist)
}
\description{
optimistic loading can lead to issues. All MSAs
should have identical length. if not theres a problem.
}
