% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaR.R
\name{msaROutput}
\alias{msaROutput}
\title{Widget output function for use in Shiny}
\usage{
msaROutput(outputId, width = "100\%", height = "100\%")
}
\arguments{
\item{outputId}{output id}

\item{width}{width}

\item{height}{height}
}
\description{
Widget output function for use in Shiny
}
